Ifpack: More namespace cleanups
